Refactored rules packages
