+++ brvs-core/src/test/java/iroha/validation/behavior/IrohaIntegrationTest.java
@@ -20,8 +20,8 @@ import iroha.protocol.QryResponses.AccountAsset;
 import iroha.protocol.TransactionOuterClass;
 import iroha.validation.config.ValidationServiceContext;
 import iroha.validation.listener.BrvsIrohaChainListener;
-import iroha.validation.rules.impl.SampleRule;
-import iroha.validation.rules.impl.TransferTxVolumeRule;
+import iroha.validation.rules.impl.core.SampleRule;
+import iroha.validation.rules.impl.assets.TransferTxVolumeRule;
 import iroha.validation.service.ValidationService;
 import iroha.validation.service.impl.ValidationServiceImpl;
 import iroha.validation.transactions.provider.impl.AccountManager;
