+++ brvs-rules/src/main/java/iroha/validation/rules/impl/core/MinimumSignatoriesAmountRule.java
@@ -3,7 +3,7 @@
  *  SPDX-License-Identifier: Apache-2.0
  */
 
-package iroha.validation.rules.impl;
+package iroha.validation.rules.impl.core;
 
 import iroha.protocol.Commands.Command;
 import iroha.protocol.Commands.RemoveSignatory;
