+++ brvs-rules/src/main/java/iroha/validation/rules/impl/core/NewBrvsRule.java
@@ -3,7 +3,7 @@
  *  SPDX-License-Identifier: Apache-2.0
  */
 
-package iroha.validation.rules.impl;
+package iroha.validation.rules.impl.core;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
