+++ brvs-rules/src/test/java/iroha/validation/rules/impl/RulesTest.java
@@ -16,6 +16,10 @@ import iroha.protocol.Commands.TransferAsset;
 import iroha.protocol.TransactionOuterClass.Transaction;
 import iroha.validation.rules.Rule;
 import iroha.validation.rules.impl.billing.BillingRule;
+import iroha.validation.rules.impl.core.MinimumSignatoriesAmountRule;
+import iroha.validation.rules.impl.core.RestrictedKeysRule;
+import iroha.validation.rules.impl.core.SampleRule;
+import iroha.validation.rules.impl.assets.TransferTxVolumeRule;
 import iroha.validation.verdict.Verdict;
 import java.io.IOException;
 import java.math.BigDecimal;
